+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [Unreleased - 0.5.1]
+## [Unreleased] - 0.5.1
 ### Added
 
 ### Changed